requirements.txt
python-telegram-bot ~= 11.1
picamera
Pillow
events
RPi.GPIO
cloudinary
watchdog
twilio ~= 6.25
Flask ~= 1.0